requirements.txt
alembic==1.0.0
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
blinker==1.4
celery==4.2.1
cffi==1.11.5
click==6.7
cryptography==2.3.1
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-MySQLdb==0.2.0
flask-paginate==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
monotonic==1.5
MySQL-python==1.2.5
mysqlclient==1.3.13
Pillow==5.3.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
python-memcached==1.59
pytz==2018.6
redis==2.10.6
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.11
vine==1.1.4
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
Wrapper==1.1.0b1
WTForms==2.2.1
